+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/pyaud-plugins/compare/v0.13.1...HEAD)
 ------------------------------------------------------------------------
+### Fixed
+- Fix version constraints for `pyaud`
 
 [0.13.1](https://github.com/jshwi/pyaud-plugins/releases/tag/v0.13.1) - 2023-02-20
 ------------------------------------------------------------------------